// functions/src/shared/enums/api-category.ts
export enum ApiCategory {
  AI = 'AI',
  Analytics = 'Analytics',
  Database = 'Database',
  Development = 'Development',
  DevOps = 'DevOps',
  Finance = 'Finance',
  Marketing = 'Marketing',
  Messaging = 'Messaging',
  Other = 'Other',
  Payment = 'Payment',
  Search = 'Search',
  Security = 'Security',
  Shipping = 'Shipping',
  Social = 'Social',
  Utilities = 'Utilities',
  Web3 = 'Web3',
}